refactoring aquamentus classes
