+++ cse3902/Sprites/EnemySprites/AquamentusSprite.cs
@@ -46,7 +46,6 @@ namespace cse3902.Sprites.EnemySprites
             spriteTexture = texture;
             remainingDelay = delay;
 
-            LoadFireballs();
             
             totalFrames = rows * columns;
             currentFrame = 0;
@@ -59,7 +58,7 @@ namespace cse3902.Sprites.EnemySprites
             this.startingPosition = startingPosition;
             center = startingPosition;
 
-            isAttacking = false;
+            isAttacking = true;
 
             DistributeFrames(columns);
 
@@ -77,37 +76,7 @@ namespace cse3902.Sprites.EnemySprites
             
         }
 
-        private void LoadFireballs()
-        {
-
-            //TODO: tweak this to position the fireballs at the mouth of aquamentus
-
-            
-
-            Vector2 startingPosition = new Vector2(0.0f, 0.0f);
-
-            //get a normalized direction vectors for the fireballs
-
-            Vector2 direction1 = Vector2.Normalize(new Vector2(-1.0f, .5f));
-            Vector2 direction2 = Vector2.Normalize(new Vector2(-1.0f, .0f));
-            Vector2 direction3 = Vector2.Normalize(new Vector2(-1.0f, -.5f));
-
-            if (this.StartingFrameIndex == (int)AquamentusSprite.FrameIndex.RightFacing)
-            {
-                direction1 = Vector2.Normalize(new Vector2(1.0f, .5f));
-                direction2 = Vector2.Normalize(new Vector2(1.0f, .0f));
-                direction3 = Vector2.Normalize(new Vector2(1.0f, -.5f));
-
-                //starting position will also need to be changed
-            }
-
-            fireball1 = new FireballSprite(spriteBatch, spriteTexture, startingPosition, direction1);
-            fireball2 = new FireballSprite(spriteBatch, spriteTexture, startingPosition, direction2);
-            fireball3 = new FireballSprite(spriteBatch, spriteTexture, startingPosition, direction3);
-        }
-
-       
-
+        
         public void Draw()
         {
             
@@ -117,12 +86,6 @@ namespace cse3902.Sprites.EnemySprites
             spriteBatch.Draw(spriteTexture, Destination, frames[currentFrame], Color.White);
             spriteBatch.End();
 
-            if (isAttacking)
-            {
-                fireball1.Draw();
-                fireball2.Draw();
-                fireball3.Draw();
-            }
         }
 
         public void Erase()
@@ -145,13 +108,7 @@ namespace cse3902.Sprites.EnemySprites
                 remainingDelay = delay;
             }
 
-            if (isAttacking)
-            {
-                fireball1.Update(gameTime, animationCompleteCallback);
-                fireball2.Update(gameTime, animationCompleteCallback);
-                fireball3.Update(gameTime, animationCompleteCallback);
 
-            }
         }
 
         // I question the need for this vector
